add detail/meta.hpp
